requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pandas==2.3.3
python-dateutil==2.9.0.post0
python-multipart==0.0.9
openpyxl==3.1.5
typer==0.12.5
pytest==8.3.2
pytest-asyncio==0.23.7
httpx==0.27.0
pydantic-settings==2.7.1

